добавлено меню выбора и запись в БД sqlite3

+++ sax_parser.py
@@ -5,15 +5,17 @@
 import xml.sax
 import os, os.path
 import sys
+import modsql3 as lite
 
 D = {}
+List=[]
 forums = 'forums.csv'
 k = 0
 seq=[1,2,4,8,9,10,11,18,19,20,
      22,23,24,25,26,27,28,29,
      31,33,34,35,36]
 
-class TorHandler( xml.sax.ContentHandler ):
+class TorHandler(xml.sax.ContentHandler):
     def __init__(self):
         self.tid = ''
         self.reg_date = ''
@@ -28,7 +30,7 @@ class TorHandler( xml.sax.ContentHandler ):
    # Call when an element starts
     def startElement(self, tag, attributes):
         self.CurrentData = tag
-        global tid,reg_date,b_size,forum_id
+        global tid, reg_date, b_size, forum_id
         if tag == 'torrent':
             tid = attributes['id']
             reg_date = attributes['registred_at']
@@ -38,7 +40,7 @@ class TorHandler( xml.sax.ContentHandler ):
                    
    # Call when an elements ends 
     def endElement(self, tag):
-        global magnet,forum,title,k
+        global magnet, forum, title, k, contents
         if self.CurrentData == 'url':
             pass
         elif self.CurrentData == 'magnet':
@@ -50,9 +52,10 @@ class TorHandler( xml.sax.ContentHandler ):
             forum = self.forum
             self.forum = ''
         elif tag == 'content':
-            #F=open('u'+str(tid)+'.txt','w',encoding='utf-8')
-            #F.write(self.contents)
+            #F=open('u'+str(tid)+'.txt','w',encoding='utf-8')   # можно сделать запись
+            #F.write(self.contents)                             # контента в файлы
             #F.close()
+            contents = self.contents
             self.contents = ''
         elif tag == 'title':
             title = self.title
@@ -60,8 +63,18 @@ class TorHandler( xml.sax.ContentHandler ):
         elif tag == 'torrent':
             k += 1
             sline = '"%s";"%s";"%s";"%s";"%s";"%s";"%s"\n' % (forum_id,forum,tid,magnet,title,b_size,reg_date)
-            fileWrite(sline)
-            if k == 1000: print(tid); k = 0
+            n = int(D.get(forum_id,0))
+            if un == '1':
+                fileWrite(sline,n)
+            else:
+                lite.check_podr(forum_id,forum)
+                lite.ins_tor(n,forum_id,tid,magnet,title,b_size,reg_date)
+                if un == '3':
+                    lite.ins_content(tid,contents)
+            if k == 1000:
+                if un != '1': lite.dbc()
+                print(tid)
+                k = 0
         self.CurrentData = ''
         
    # Call when a character is read
@@ -92,8 +105,8 @@ def fileClose():
         globals()[s].close()
 
 
-def fileWrite(stroka):
-    n = int(D.get(forum_id,0)) 
+def fileWrite(stroka,n):
+     
     if not sys.version[0] == '3':
         st=u''
         st=stroka
@@ -111,37 +124,41 @@ def instor(): # insert start and end tags '<torrents>'
     FB.seek(0)
     FB.write(b'<?xml version="1.0"?>\n<torrents>      ')
     FB.close()
-    FB = open(backup,'ab')
-    FB.write(b'</torrents>')
+    FB = open(backup,'a',encoding='utf-8')
+    FB.write('</torrents>')
     FB.close()
 
 def load_forums3():
     # Load dic FORUMS vers 3
     for line in open(forums, encoding = 'utf-8'):
-        forum = line.split(sep=';"')[0]
-        category = line.split(sep='";')[1]
+        L=line.split(sep=';')
+        forum = L[0]
+        name_forum=L[1][1:-1]
+        category = L[-1]
         D[forum] = category
-    print('Dictionary loaded')
-
+        List.append((int(forum),name_forum,int(category)))
+        
 def load_forums2():
     # Load dic FORUMS vers 2
     for line in open(forums,'rb').xreadlines():
         xline=line.decode('utf-8')
-        forum = xline.split(';"')[0]
-        category = xline.split('";')[1]
+        L=xline.split(sep=';')
+        forum = L[0]
+        name_forum=L[1][1:-1]
+        category = L[-1]
         D[forum] = category
-    print('Dictionary loaded')
+        List.append((int(forum),name_forum,int(category)))
     
-def test_tor():
-    # test for '<torrents>'
+def test_tor():                 # test for '<torrents>'
     f = open(backup,'rb')
     a = str(f.read(50))
     f.close()
     if a.find('<torrents>') == -1: instor()
 
-    
 # START PROG ->
 if __name__ == '__main__':
+    global un
+    
     for f in os.listdir('.'):
         if f[:8] == 'backup.2':
             backup = f
@@ -149,29 +166,51 @@ if __name__ == '__main__':
     	
     period = backup.split(".")[1][:8]
     catalog = './'+period
-    if os.path.exists(catalog):
-        for f in os.listdir(catalog):
-            os.remove(catalog+'/'+f)
-    else:
-        os.mkdir(catalog)
-    
-    test_tor()
+    dirDB = './DB'
+    print('Обрабатываемый файл: ' + backup)
 
-    if sys.version[0] == '3':
-        load_forums3()
-    else:
-        load_forums2()
-    
-    # create an XMLReader
-    parser = xml.sax.make_parser()
-    # turn off namepsaces
-    parser.setFeature(xml.sax.handler.feature_namespaces, 0)
-    
-    # override the default ContextHandler
-    Handler = TorHandler()
-    parser.setContentHandler( Handler )
+    un = input('''Выберите нужное действие:
+    1. Сохранить в CSV;
+    2. Сохранить в БД(sqlite);
+    3. Сохранить описания раздач в отдельную БД(дополнительно, со сжатием);
+    4. Выход (любой ввод кроме 1,2,3)\n''')
 
-    fileOpen()
-    parser.parse(backup)
-    fileClose()
-    print('Extract OK!')
+    if un in ('1','2','3'):
+        if sys.version[0] == '3':
+            load_forums3()
+        else:
+            load_forums2()
+        print('Dictionary loaded')
+        
+        if un == '1':
+            if os.path.exists(catalog):
+                for f in os.listdir(catalog):
+                    os.remove(catalog+'/'+f)
+            else:
+                os.mkdir(catalog)
+        elif un == '2' or '3':
+            if not os.path.exists(dirDB):
+                os.mkdir(dirDB)
+
+            lite.create_db(dirDB+'/')
+            lite.ins_forums(List)
+            if un == '3':
+                lite.create_db_content(dirDB+'/')
+     
+        
+
+        parser = xml.sax.make_parser()                              # create an XMLReader
+        parser.setFeature(xml.sax.handler.feature_namespaces, 0)    # turn off namepsaces
+        Handler = TorHandler()                                      # override the default ContextHandler
+        parser.setContentHandler( Handler )
+
+        if un == '1': fileOpen()
+        parser.parse(backup)
+        if un == '1':
+            fileClose()
+        else:
+            lite.close_db()
+        print('Extract OK!')
+
+    else:
+        print('Goodbye!')
